+++ CHANGELOG.md
@@ -50,6 +50,8 @@
 
 * Return `SWIFT_NAME` when generating Objective-C docs.  
   [Ibrahim Ulukaya](https://github.com/ulukaya)
+* libsourcekitdInProc search strategy improved on Linux. Supports swiftenv.  
+  [Alexander Lash](https://github.com/abl)
 
 ##### Bug Fixes